COL-1857: Add missing conversation icons
Added missing marketplace, offer and private icons for the
ts-document-card to be able to support the design of Collaboration
Messages app.

Because IE11 doesn’t have a Shadow DOM, the polyfill adds to the
encapsulated styles of the component the “unique” css class which is has
same name as the component name, and in this case it was ts-table.
Because of this the .ts-table rule from the old ts-ui css affected the
view of the table.