Move peer deps to dependencies
Based on the changes that have been done about peerDependencies in the npm@7 it caused error for projects that've being using the radio-group or tabs. `tab` and `radio` components are added to dependencies of `tabs` and `radio-group` since they are actually useless without those components and it makes developers life easier to just handle one dep if they only need those group components and not import the child component.