feat(select): add custom filtering to select component
Now it's possible to do custom filtering beside the basic filter
that is done by default for item.title text.
Now developers can listen to `filter-value-change` event,
to get the user inputed value for search/filter and
pass `filteredItems` to override the default internal
filter/search through the already provided items.
To make it possible to do the filtering asynchronously, added
loading state to the select component.

Note: Items that are provided for `filteredItems`, should be also
included in `items`, which means developers should make sure they
add them to items list if they are not already there and not remove
other items for multi select, since they can be already in current
selection state.