fix(select): do not set input value for multi select and no apply
In the no apply button mode, since we change the selection without
changing the `opened` state of the component, it would set the input
value, which was overriding the entered filter value and causing bad
user experience.
Since there's no scenario that we want to set the input value while
select-menu is open, just ignoring the update input value request.